Fixed Bugfs

+++ app/controllers/api/inventory_requests_controller.rb
@@ -9,7 +9,7 @@ class Api::InventoryRequestsController < ApplicationController
 
   def create
     if @inventory
-      if @inventory.quantity.to_f < params[:quantity_asked].to_i
+      if @inventory.quantity.to_f < params[:quantity_asked].to_f
         render json: @inventory, status: 403
       else
         @inventory.request_for_inventory(current_employee, params[:quantity_asked], params[:date_of_use])
